--- compiled/index.css
@@ -1,143 +1,0 @@
-:root {
-  color-scheme: dark;
-}
-
-* {
-  box-sizing: border-box;
-}
-
-._palette-module__backdrop__1c4BQ {
-  display: flex;
-  overflow: hidden;
-  position: fixed;
-  top: 0;
-  left: 0;
-  z-index: 50;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  width: 100vw;
-  height: 100vh;
-  background: #00000090;
-  -webkit-backdrop-filter: blur(2px);
-  backdrop-filter: blur(2px);
-  -webkit-user-select: none;
-  -moz-user-select: none;
-  user-select: none;
-}
-
-._palette-module__dialog__1zn65 {
-  padding: 0.5rem;
-  background-color: #181818;
-  width: 30%;
-  border-radius: 1rem;
-  border: solid 1px #3f3f3f;
-}
-
-._palette-module__search__3vbGc {
-  color: white;
-  border: none;
-  padding: 1.25rem;
-  margin-bottom: 0.5rem;
-  background-color: transparent;
-  font-size: 1rem;
-  line-height: 1.5rem;
-  width: 100%;
-  border-bottom: #777777 1px solid;
-  outline: 0;
-}
-
-._palette-module__wrapper__SKV7e {
-  display: flex;
-  overflow-y: scroll;
-  flex-direction: column;
-  width: 100%;
-  height: -webkit-fit-content;
-  height: -moz-fit-content;
-  height: fit-content;
-  max-height: 320px;
-  transition: height 0.1s ease 0s;
-  will-change: height;
-}
-
-._palette-module__no_results__29zn5 {
-  display: flex;
-  margin-top: 2.5rem;
-  margin-bottom: 2.5rem;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  height: 100%;
-}
-
-._palette-module__description__18LcE {
-  margin-top: 0.5rem;
-  font-size: 1.125rem;
-  line-height: 1.75rem;
-}
-
-._palette-module__command__30iJb {
-  display: flex;
-  position: relative;
-  margin-top: 0.25rem;
-  margin-bottom: 0.25rem;
-  color: #d1d5db;
-  font-size: 1.125rem;
-  line-height: 1.75rem;
-  align-items: center;
-  text-decoration: none;
-  cursor: pointer;
-}
-
-._palette-module__command__30iJb > svg {
-  margin: 0 10px 0 15px;
-  width: 20px;
-  height: 20px;
-}
-
-._palette-module__title__3WYB3 {
-  margin: 10px 0 15px 10px;
-  font-size: 14px;
-  color: #777777;
-}
-
-._palette-module__selected__3EiZZ {
-  position: absolute;
-  width: 100%;
-  height: 4rem;
-  border-radius: 0.5rem;
-  background: #ffffff20;
-}
-
-._palette-module__text__O2N7u {
-  width: 90%;
-  margin: 1rem 0;
-  font-size: 1rem;
-  line-height: 1.5rem;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-@media only screen and (max-width: 1440px) {
-  ._palette-module__dialog__1zn65 {
-    width: 40%;
-  }
-}
-
-@media only screen and (max-width: 1024px) {
-  ._palette-module__dialog__1zn65 {
-    width: 50%;
-  }
-}
-
-@media only screen and (max-width: 768px) {
-  ._palette-module__dialog__1zn65 {
-    width: 70%;
-  }
-}
-
-@media only screen and (max-width: 600px) {
-  ._palette-module__dialog__1zn65 {
-    width: 90%;
-  }
-}
