+++ src/utils/parse.ts
@@ -1,15 +1,5 @@
-import { CategoryCommand, ParseProps } from '../types'
-
-/* Function used for running the command */
-export const run = (command: CategoryCommand) => {
-  /* If the command's perform is not undefined, run the command */
-  if (typeof command.perform !== 'undefined') return command.perform?.()
-  /* If the command's href is not undefined, then open a new window */ else if (
-    typeof command.href !== 'undefined'
-  )
-    /* Ternary, just checking whether or not the command should open in a new tab or not */
-    window.open(command.href, command.newTab ? '_blank' : '_self')
-}
+import { ParseProps } from '../types'
+import run from './run'
 
 /* Function used for parsing a shortcut */
 const parse = ({ command, event, map }: ParseProps) => {
